Add user and workout program

+++ src/main/kotlin/com/congen/model/User.kt
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
+import java.math.BigDecimal
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -11,6 +12,6 @@ data class User(
     @param:JsonProperty("id") val id: Int? = null,
     @param:JsonProperty("name") val name: String,
     @param:JsonProperty("age") val age: Int,
-    @param:JsonProperty("height") val height: Double,
-    @param:JsonProperty("weight") val weight: Double,
+    @param:JsonProperty("height") val height: BigDecimal,
+    @param:JsonProperty("weight") val weight: BigDecimal,
 )
